# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workout/today": {
      "get": {
        "description": "Returns the recommended workout plan for today. Returns null if workout preferences are not set.",
        "operationId": "WorkoutController_getTodayWorkout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayWorkoutResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get today's suggested workout",
        "tags": [
          "Workout"
        ]
      }
    },
    "/workout/sessions": {
      "get": {
        "description": "Returns all available workout sessions with their exercises",
        "operationId": "WorkoutController_getSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutSessionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get workout sessions",
        "tags": [
          "Workout"
        ]
      }
    },
    "/workout/exercises": {
      "get": {
        "description": "Returns a list of all available exercises from the database",
        "operationId": "WorkoutController_getExercises",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseListResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get all exercises",
        "tags": [
          "Workout"
        ]
      }
    },
    "/workout/preferences": {
      "get": {
        "description": "Returns the stored workout preferences for the authenticated user",
        "operationId": "WorkoutController_getWorkoutPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPreferencesResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get workout preferences",
        "tags": [
          "Workout"
        ]
      },
      "post": {
        "description": "Creates or updates workout preferences. All fields are optional — send only what you want to change.",
        "operationId": "WorkoutController_saveWorkoutPreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveWorkoutPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPreferencesResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Save workout preferences",
        "tags": [
          "Workout"
        ]
      }
    },
    "/workout/session": {
      "post": {
        "operationId": "WorkoutController_createSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Create a new workout session",
        "tags": [
          "Workout"
        ]
      }
    },
    "/workout/session/{id}": {
      "get": {
        "operationId": "WorkoutController_getSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get a specific workout session",
        "tags": [
          "Workout"
        ]
      },
      "patch": {
        "operationId": "WorkoutController_updateSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Update a workout session",
        "tags": [
          "Workout"
        ]
      },
      "delete": {
        "operationId": "WorkoutController_deleteSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session deleted successfully"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Delete a workout session",
        "tags": [
          "Workout"
        ]
      }
    },
    "/workout/my-sessions": {
      "get": {
        "operationId": "WorkoutController_getUserSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get all workout sessions for the authenticated user",
        "tags": [
          "Workout"
        ]
      }
    },
    "/workout/log": {
      "post": {
        "operationId": "WorkoutController_createWorkoutLog",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutLogRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutLogResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Create a new workout log",
        "tags": [
          "Workout"
        ]
      }
    },
    "/workout/log/{id}": {
      "get": {
        "operationId": "WorkoutController_getWorkoutLog",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutLogResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get a specific workout log",
        "tags": [
          "Workout"
        ]
      },
      "patch": {
        "operationId": "WorkoutController_updateWorkoutLog",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutLogRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutLogResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Update a workout log",
        "tags": [
          "Workout"
        ]
      },
      "delete": {
        "operationId": "WorkoutController_deleteWorkoutLog",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workout log deleted successfully"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Delete a workout log",
        "tags": [
          "Workout"
        ]
      }
    },
    "/workout/logs": {
      "get": {
        "operationId": "WorkoutController_getWorkoutLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutLogResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get all workout logs for the authenticated user",
        "tags": [
          "Workout"
        ]
      }
    },
...
  },
  "info": {
    "title": "nestjs-starter API",
    "description": "nestjs-starter API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "google-oauth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://localhost:3000/auth/google",
            "scopes": {
              "email profile": "Get email and profile info"
            }
          }
        }
      }
    },
    "schemas": {
...
      "SetDto": {
        "type": "object",
        "properties": {
          "kg": {
            "type": "number",
            "example": 60
          },
          "reps": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "kg",
          "reps"
        ]
      },
      "ExerciseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ex-001"
          },
          "name": {
            "type": "string",
            "example": "Bench Press"
          },
          "desc": {
            "type": "string",
            "example": "Chest, Triceps, Shoulders"
          },
          "image": {
            "type": "string",
            "example": "https://cdn.example.com/bench-press.png"
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "desc",
          "image",
          "sets"
        ]
      },
      "WorkoutSessionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "session-001"
          },
          "title": {
            "type": "string",
            "example": "Chest & Triceps"
          },
          "content": {
            "type": "string",
            "example": "Push day focused on upper body strength"
          },
          "category": {
            "type": "string",
            "example": "Strength"
          },
          "exercisesCount": {
            "type": "number",
            "example": 4
          },
          "position": {
            "type": "number",
            "example": 1
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "category",
          "exercisesCount",
          "position",
          "exercises"
        ]
      },
      "TodayWorkoutResponseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Suggested Muscle"
          },
          "highlight": {
            "type": "string",
            "example": "Bicep & Back"
          },
          "subtext": {
            "type": "string",
            "example": "Optional for hypertrophy"
          },
          "duration": {
            "type": "number",
            "example": 45
          },
          "intensity": {
            "type": "string",
            "example": "hard",
            "enum": [
              "easy",
              "moderate",
              "hard"
            ]
          },
          "session": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutSessionDto"
            }
          }
        },
        "required": [
          "title",
          "highlight",
          "subtext",
          "duration",
          "intensity",
          "session"
        ]
      },
      "ExerciseListResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Bench Press"
          },
          "secondaryInfo": {
            "type": "string",
            "example": "Chest, Triceps, Shoulders"
          },
          "exerciseType": {
            "type": "string",
            "example": "Strength"
          },
          "image": {
            "type": "string",
            "example": "https://r2.heracle.fit/exercises/1.jpg"
          }
        },
        "required": [
          "id",
          "name",
          "secondaryInfo",
          "exerciseType",
          "image"
        ]
      },
      "WorkoutPreferencesResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-..."
          },
          "fitnessLevel": {
            "type": "object",
            "example": "intermediate"
          },
          "workoutFrequencyPerWeek": {
            "type": "object",
            "example": 4
          },
          "preferredWorkoutType": {
            "type": "object",
            "example": "strength"
          },
          "injuries": {
            "type": "object",
            "example": "left knee pain"
          },
          "availableDays": {
            "example": [
              "monday",
              "wednesday",
              "friday"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredWorkoutTime": {
            "type": "object",
            "example": "morning"
          },
          "sessionDurationMins": {
            "type": "object",
            "example": 45
          },
          "updatedAt": {
            "type": "object",
            "example": "2026-02-25T11:30:00.000Z"
          }
        },
        "required": [
          "id",
          "fitnessLevel",
          "workoutFrequencyPerWeek",
          "preferredWorkoutType",
          "injuries",
          "availableDays",
          "preferredWorkoutTime",
          "sessionDurationMins"
        ]
      },
      "SaveWorkoutPreferencesDto": {
        "type": "object",
        "properties": {
          "fitnessLevel": {
            "type": "string",
            "example": "intermediate",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ]
          },
          "workoutFrequencyPerWeek": {
            "type": "number",
            "example": 4,
            "description": "Number of workouts per week"
          },
          "preferredWorkoutType": {
            "type": "string",
            "example": "strength",
            "enum": [
              "strength",
              "cardio",
              "yoga",
              "hiit",
              "flexibility",
              "mixed"
            ]
          },
          "injuries": {
            "type": "string",
            "example": "left knee pain",
            "description": "Any injuries or physical limitations"
          },
          "availableDays": {
            "example": [
              "monday",
              "wednesday",
              "friday"
            ],
            "description": "Days available to work out",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredWorkoutTime": {
            "type": "string",
            "example": "morning",
            "enum": [
              "morning",
              "afternoon",
              "evening",
              "flexible"
            ]
          },
          "sessionDurationMins": {
            "type": "number",
            "example": 45,
            "description": "Preferred session duration in minutes"
          }
        }
      },
      "CreateSessionRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Morning Push Session",
            "description": "Name of the workout session"
          },
          "category": {
            "example": [
              "Strength",
              "Chest"
            ],
            "description": "Categories associated with the session",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sessionData": {
            "type": "object",
            "description": "JSON array of exercise targets",
            "example": [
              {
                "exercise id": 1,
                "set1": {
                  "targetRep": 10,
                  "targetKg": 10
                }
              }
            ]
          }
        },
        "required": [
          "name",
          "category",
          "sessionData"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Morning Push Session"
          },
          "category": {
            "example": [
              "Strength",
              "Chest"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sessionData": {
            "type": "object"
          },
          "exerciseImageBaseUrl": {
            "type": "string",
            "example": "https://r2.heracle.fit/exercises"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "sessionData",
          "exerciseImageBaseUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateSessionRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Updated Session Name"
          },
          "category": {
            "example": [
              "Strength"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sessionData": {
            "type": "object",
            "description": "JSON array of exercise targets"
          }
        }
      },
      "CreateWorkoutLogRequestDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "number",
            "example": 1,
            "description": "Optional session ID if this log is part of a session"
          },
          "logData": {
            "type": "object",
            "description": "JSON array of exercise performance data",
            "example": [
              {
                "exercise id": 1,
                "set1": {
                  "Rep": 10,
                  "Kg": 10
                },
                "note": "Felt good"
              }
            ]
          },
          "notes": {
            "type": "string",
            "example": "Great workout today!",
            "description": "General notes for the workout"
          }
        },
        "required": [
          "logData"
        ]
      },
      "WorkoutLogResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "userId": {
            "type": "string",
            "example": "uuid-..."
          },
          "sessionId": {
            "type": "object",
            "example": 1
          },
          "logData": {
            "type": "object"
          },
          "notes": {
            "type": "object"
          },
          "exerciseImageBaseUrl": {
            "type": "string",
            "example": "https://r2.heracle.fit/exercises"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "logData",
          "exerciseImageBaseUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateWorkoutLogRequestDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "number",
            "example": 1
          },
          "logData": {
            "type": "object",
            "description": "JSON array of exercise performance data",
            "example": [
              {
                "exercise id": 1,
                "set1": {
                  "Rep": 12,
                  "Kg": 12
                }
              }
            ]
          },
          "notes": {
            "type": "string",
            "example": "Updated notes"
          }
        }
      },
...
    }
  }
}